+++ packages/bridge/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './bridger'
 export * from './config'
+export * from './internal'
 export * from './state'
 export * from './token'
 export * from './tx'