+++ packages/nft-checkout/src/const/tokens.ts
@@ -3,6 +3,30 @@ import { ChainNames, EVM_CHAINS } from '@rarimo/shared'
 
 const chapelChain = EVM_CHAINS.find(i => i.name === ChainNames.Chapel)!
 const fujiChain = EVM_CHAINS.find(i => i.name === ChainNames.Fuji)!
+const ethChain = EVM_CHAINS.find(i => i.name === ChainNames.Ethereum)!
+const bscChain = EVM_CHAINS.find(i => i.name === ChainNames.BinanceSmartChain)!
+
+export const ETH_TOKEN_LIST: Token[] = [
+  newToken(
+    ethChain,
+    '0xb8c77482e45f1f44de1745f52c74426c631bdd52',
+    'Binance Coin',
+    'BNB',
+    18,
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/bnb-logo.png',
+  ),
+]
+
+export const BINANCE_TOKEN_LIST: Token[] = [
+  newToken(
+    bscChain,
+    '0xb8c77482e45f1f44de1745f52c74426c631bdd52',
+    'USD Coin',
+    'USDC',
+    18,
+    'https://raw.githubusercontent.com/rarimo/js-sdk/main/assets/logos/usdc-logo.png',
+  ),
+]
 
 export const PANCAKE_SWAP_TESTNET_TOKEN_LIST: Token[] = [
   newToken(