+++ CHANGELOG.md
@@ -21,6 +21,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `all` - Node resolution in the `package.json` files
 
+## [1.5.0-rc.16] - 2023-05-23
+### Changed
+- `@rarimo/nft-checkout` - Extend token list for Ethereum, Binance Smart Chain to resolve target token by address got from backend
+
+## [1.5.0-rc.15] - 2023-05-23
+### Fixed
+- `@rarimo/nft-checkout` - Checking decimals for non-zero for the payment token
+
+## [1.5.0-rc.14] - 2023-05-22
+### Added
+- `@rarimo/nft-checkout` - Checking if payment token has liquidity pair with the target token
+
+## [1.5.0-rc.13] - 2023-05-22
+### Fixed
+- `@rarimo/react-provider` - Calling clear handlers on non-existent provider
+- `@rarimo/provider` - Extended `RuntimeEнфкrror` with `orginalError` field
+
+## [1.5.0-rc.12] - 2023-05-17
+### Fixed
+- `@rarimo/nft-checkout`: Parsing of the balance received after multicall
+
 ## [1.5.0-rc.11] - 2023-05-16
 ### Added
 - `@rarimo/react-nft-checkout`: getDestinationTx method into `DappContextProvider`
@@ -218,7 +239,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.11...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.16...HEAD
+[1.5.0-rc.16]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.15...1.5.0-rc.16
+[1.5.0-rc.15]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.14...1.5.0-rc.15
+[1.5.0-rc.14]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.13...1.5.0-rc.14
+[1.5.0-rc.13]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.12...1.5.0-rc.13
+[1.5.0-rc.12]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.11...1.5.0-rc.12
 [1.5.0-rc.11]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.10...1.5.0-rc.11
 [1.5.0-rc.10]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.9...1.5.0-rc.10
 [1.5.0-rc.9]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.8...1.5.0-rc.9